requirements.txt
# This file includes all requirements including
# those that are optional.
pandas
numpy
blinker
matplotlib
numexpr
click
xarray
scipy
seaborn
pydataset
h5py
pydons
tabulate
sqlalchemy
qtpy
feather-format
clevercsv
typer
